+++ tests/mocks/__init__.py
@@ -162,6 +162,7 @@ def build_guild(offset: int = 1) -> discord.Guild:
     guild.owner_id = OWNER_USER_ID
     guild.create_category_channel = AsyncMock()
     guild.bitrate_limit = 64000.0
+    guild.preferred_locale = discord.Locale.american_english
     return guild
 
 
@@ -261,6 +262,7 @@ def mock_discord_guild(guild: Guild) -> discord.Guild:
     discord_guild.id = guild.xid
     discord_guild.name = guild.name
     discord_guild.bitrate_limit = 64000.0
+    discord_guild.preferred_locale = discord.Locale.american_english
     return discord_guild